Unidad2/Clase11_Harold_vergara/back/project/serializers/computador_serializer.py
from rest_framework import serializers
from api.models.computador import Computador

class ComputadorSerializers(serializers.ModelSerializer):
    class Meta:
        model = Computador 
        exclude = ["id"]
