Unidad2/Clase11_Harold_vergara/back/requirements.txt
Django==3.0.8
djangorestframework==3.11.0
django-model-utils==4.0.0
